requirements.txt
playwright
python-dotenv
streamlit
